requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.10.0
confusable-homoglyphs==3.2.0
Django==3.2.7
django-bootstrap4==3.0.1
django-registration==3.2
django-tinymce==3.3.0
Pillow==8.3.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
